+++ src/pages/Admin/Applicants/Filter/interfaces.ts
@@ -1,5 +1,5 @@
 import { IUseApplicantsFilter } from "$hooks";
-import { ApplicantsFilter } from "$models/ApplicantsFilter";
+import { ApplicantsFilter } from "$models/SearchFilters/ApplicantsFilter";
 import { ApplicantType } from "$interfaces/Applicant";
 import { ICareer } from "$interfaces/Career";
 import { FormikHelpers } from "formik";
